Synrgi Navigation Patch
Two patches to enable the navigation bar toggle. More details in each
patch. Currently requires reboot to toggle.

_FrBase.patch -> frameworks/base
_Settings.patch -> packages/apps/Settings

Also, rom must be built with Navigation Bar config value set to false
in order to enable toggle. This is to prevent user from toggling off
navigation bar on a device that does not have hardware keys, which
would require a wipe to fix.